+++ docs/NEWS.rst
@@ -8,6 +8,13 @@ of XIST. For a description of how to update your code to each versions of XIST
 see :ref:`MIGRATION`.
 
 
+Changes in 5.50 (released 07/??/2019)
+-------------------------------------
+
+*	There's a new option ``-a``/``--ascii`` for running PySQL scripts: With this
+	PySQL will not use unicode characters for drawing fancy boxes.
+
+
 Changes in 5.49 (released 07/04/2019)
 -------------------------------------